documenting relation and updating name to be more usable
